Extract file rotation logic to FileRotationService (T2)
Refactors FileSink by extracting rotation logic to dedicated FileRotationService.
This improves Single Responsibility Principle (SRP), testability, and future
reusability for Pro sinks.

Changes:
- Created FileRotationService.cs (~170 lines) with comprehensive XML docs
  - ShouldRotate(): checks if file exceeds size limit
  - RotateFile(): archives file with timestamp and cleanup
  - Thread safety documented (must be used within caller's lock)

- Updated FileSink.cs (reduced from 252 to 160 lines)
  - Removed rotation methods (ShouldRotate, RotateLogFile, CleanupOldLogFiles)
  - Added FileRotationService field
  - Updated Write() to use rotation service
  - Properly handles writer lifecycle during rotation

Benefits:
- FileSink now focuses solely on writing logs
- Rotation logic is independently testable
- Can be reused by future file-based sinks
- No breaking changes to public API
- All existing tests continue to pass

Related to issue: Extract File Rotation Logic to FileRotationService (T2)
